[Style] 가독성을 위한 코드 수정
후행클로저로 변경, 파일 분리, 개행문자 삽입
